+++ tests/testthat/test-anova_stats.R
@@ -32,4 +32,32 @@ if (require("testthat") && require("sjstats")) {
     anova_stats(car::Anova(m, type = 2))
     anova_stats(car::Anova(m, type = 3))
   })
+
+
+  set.seed(123)
+  fit <- aov(
+    c12hour ~ as.factor(e42dep) + as.factor(c172code) + c160age,
+    data = efc
+  )
+
+  test_that("omega_sq", {
+    omega_sq(fit, partial = TRUE, ci.lvl = 0.95, n = 50)
+  })
+
+  set.seed(123)
+  data(mtcars)
+  m <-
+    stats::aov(
+      formula = mpg ~ wt + qsec + Error(disp / am),
+      data = mtcars
+    )
+
+  test_that("anova_stats", {
+    anova_stats(m, digits = 3)
+    anova_stats(m, digits = 5)
+    eta_sq(m, partial = TRUE, ci.lvl = 0.95, n = 10)
+    eta_sq(m, partial = FALSE, ci.lvl = 0.95, n = 10)
+    omega_sq(m, partial = TRUE, ci.lvl = 0.95, n = 10)
+    omega_sq(m, partial = FALSE, ci.lvl = 0.95, n = 10)
+  })
 }